etl/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic>=1.10
pytest
httpx
